refactor(console): extract shared router mock for chat tests
Two chat test files were duplicating ~35 lines of useSyncExternalStore-
backed router mock scaffolding with real but small divergences (nullable
vs non-null initial session, lastTo/lastReplace tracking). Fold into a
single createRouterMock factory under __test-utils/ that takes an
initial session id and returns the mock module shape.
